# [O11Y-197] optimize min/max algorithms on x86 with AVX-512 support