--- a/internal/bits/bits_amd64.go
+++ b/internal/bits/bits_amd64.go
@@ -1,0 +1,7 @@
+package bits
+
+import "golang.org/x/sys/cpu"
+
+var hasAVX512 = cpu.X86.HasAVX512 &&
+	cpu.X86.HasAVX512F &&
+	cpu.X86.HasAVX512VL