--- a/internal/bits/bits_amd64.go
+++ b/internal/bits/bits_amd64.go
@@ -5,3 +5,6 @@ import "golang.org/x/sys/cpu"
 var hasAVX512 = cpu.X86.HasAVX512 &&
 	cpu.X86.HasAVX512F &&
 	cpu.X86.HasAVX512VL
+
+var hasAVX512MinMaxBool = hasAVX512 &&
+	cpu.X86.HasAVX512VPOPCNTDQ