--- a/internal/bits/count.go
+++ b/internal/bits/count.go
@@ -1,0 +1,3 @@
+package bits
+
+func CountByte(data []byte, value byte) int { return countByte(data, value) }