--- a/internal/bits/count_default.go
+++ b/internal/bits/count_default.go
@@ -1,0 +1,9 @@
+//go:build purego || !amd64
+
+package bits
+
+import "bytes"
+
+func countByte(data []byte, value byte) int {
+	return bytes.Count(data, []byte{value})
+}