--- a/internal/bits/max.go
+++ b/internal/bits/max.go
@@ -2,89 +2,28 @@ package bits
 
 import "bytes"
 
-func MaxBool(data []bool) (max bool) {
-	return len(data) > 0 && !boolEqualAll(data, false)
-}
+func MaxBool(data []bool) (max bool) { return maxBool(data) }
 
-func MaxInt32(data []int32) (max int32) {
-	if len(data) > 0 {
-		max = data[0]
+func MaxInt32(data []int32) (max int32) { return maxInt32(data) }
 
-		for _, value := range data {
-			if value > max {
-				max = value
-			}
-		}
-	}
-	return max
-}
+func MaxInt64(data []int64) (max int64) { return maxInt64(data) }
 
-func MaxInt64(data []int64) (max int64) {
-	if len(data) > 0 {
-		max = data[0]
+func MaxUint32(data []uint32) (max uint32) { return maxUint32(data) }
 
-		for _, value := range data {
-			if value > max {
-				max = value
-			}
-		}
-	}
-	return max
-}
+func MaxUint64(data []uint64) (max uint64) { return maxUint64(data) }
 
-func MaxUint32(data []uint32) (max uint32) {
-	if len(data) > 0 {
-		max = data[0]
+func MaxFloat32(data []float32) (max float32) { return maxFloat32(data) }
 
-		for _, value := range data {
-			if value > max {
-				max = value
-			}
-		}
-	}
-	return max
-}
-
-func MaxUint64(data []uint64) (max uint64) {
-	if len(data) > 0 {
-		max = data[0]
-
-		for _, value := range data {
-			if value > max {
-				max = value
-			}
-		}
-	}
-	return max
-}
-
-func MaxFloat32(data []float32) (max float32) {
-	if len(data) > 0 {
-		max = data[0]
-
-		for _, value := range data {
-			if value > max {
-				max = value
-			}
-		}
-	}
-	return max
-}
-
-func MaxFloat64(data []float64) (max float64) {
-	if len(data) > 0 {
-		max = data[0]
-
-		for _, value := range data {
-			if value > max {
-				max = value
-			}
-		}
-	}
-	return max
-}
+func MaxFloat64(data []float64) (max float64) { return maxFloat64(data) }
 
 func MaxFixedLenByteArray(size int, data []byte) (max []byte) {
+	if size == 16 {
+		return maxBE128(data)
+	}
+	return maxFixedLenByteArray(size, data)
+}
+
+func maxFixedLenByteArray(size int, data []byte) (max []byte) {
 	if len(data) > 0 {
 		max = data[:size]