--- a/internal/bits/max_amd64.h
+++ b/internal/bits/max_amd64.h
@@ -21,7 +21,7 @@
 //
 //   A[1] > B[1] || (A[1] == B[1] && A[0] > B[0])
 //
-// Values in the source vector that evalute to true on this expression are
+// Values in the source vector that evaluate to true on this expression are
 // written to the vector of maximum values, and their indexes are written to
 // the vector of indexes.
 #define vpmaxu128(srcValues, srcIndexes, maxValues, maxIndexes, K1, K2, R1, R2, R3, M) \