--- a/internal/bits/max_amd64.s
+++ b/internal/bits/max_amd64.s
@@ -570,7 +570,7 @@ loop16:
     // Unless the input was aligned on 256 bytes, we need to perform a few more
     // iterations on the remaining elements.
     //
-    // This loop is also take if the CPU has no support for AVX-512.
+    // This loop is also taken if the CPU has no support for AVX-512.
 loop:
     MOVBEQQ (AX), R8
     MOVBEQQ (BX), R9