--- a/internal/bits/max_default.go
+++ b/internal/bits/max_default.go
@@ -98,7 +98,7 @@ func maxBE128(data []byte) (min []byte) {
 				m, j = x, i
 			case x == m:
 				y := binary.BigEndian.Uint64(be128[i][8:])
-				n := binary.BigEndian.Uint64(be128[j][:8])
+				n := binary.BigEndian.Uint64(be128[j][8:])
 				if y > n {
 					m, j = x, i
 				}