--- a/internal/bits/min_amd64.go
+++ b/internal/bits/min_amd64.go
@@ -19,3 +19,6 @@ func minFloat32(data []float32) float32
 
 //go:noescape
 func minFloat64(data []float64) float64
+
+// go:noescape
+func minBE128(data []byte) []byte