… internal/bits/bits_test.go
… internal/bits/count_test.go
… internal/bits/max_test.go
… internal/bits/min_amd64.h
… internal/bits/min_amd64.s
… internal/bits/min_default.go
… internal/bits/min_test.go
… internal/bits/minmax.go
… internal/bits/minmax_amd64.go
… internal/bits/minmax_amd64.s
… internal/bits/minmax_default.go
… internal/bits/minmax_test.go
… internal/bits/unsafe.go
… page.go