I've started this a couple weeks ago but didn't find the time to wrap up the work until today, so just submitting now.

This PR revisits the implementations of a couple of internal algorithms that we use when creating parquet files, using vectorized AVX-512 instructions to improve the throughput of those operations on parquet columns.

The assembly code can be complex in places, but the work being done is very simple in general, we look for min/max values in an arrays of primitive types. I've made sure to invest in thorough testing to minimize the risk of seeing regressions introduced by those optimizations.

For what it's worth, those are the measured improvements on some of the benchmarks I added in the PR:

```
name                             old time/op    new time/op      delta
MinMaxBool/4KiB                    1.23µs ± 0%      0.04µs ± 0%    -96.79%  (p=0.000 n=8+10)
MinMaxBool/256KiB                  77.6µs ± 0%       3.3µs ± 0%    -95.73%  (p=0.000 n=10+9)
MinMaxBool/2048KiB                  624µs ± 0%        64µs ± 8%    -89.78%  (p=0.000 n=9+10)
MinMaxInt32/4KiB                    640ns ± 0%        44ns ± 0%    -93.11%  (p=0.000 n=9+10)
MinMaxInt32/256KiB                 40.7µs ± 0%       3.3µs ± 0%    -91.84%  (p=0.000 n=10+9)
MinMaxInt32/2048KiB                 327µs ± 0%        60µs ± 7%    -81.57%  (p=0.000 n=10+10)
MinMaxInt64/4KiB                    331ns ± 0%        45ns ± 0%    -86.40%  (p=0.000 n=9+8)
MinMaxInt64/256KiB                 20.9µs ± 0%       3.4µs ± 0%    -83.87%  (p=0.000 n=10+9)
MinMaxInt64/2048KiB                 169µs ± 0%        62µs ± 8%    -63.28%  (p=0.000 n=9+10)
MinMaxUint32/4KiB                   905ns ± 0%        44ns ± 0%    -95.14%  (p=0.000 n=9+8)
MinMaxUint32/256KiB                58.1µs ± 0%       3.3µs ± 0%    -94.31%  (p=0.000 n=10+9)
MinMaxUint32/2048KiB                467µs ± 0%        59µs ± 8%    -87.32%  (p=0.000 n=10+10)
MinMaxUint64/4KiB                   450ns ± 0%        46ns ± 0%    -89.81%  (p=0.000 n=10+8)
MinMaxUint64/256KiB                29.1µs ± 0%       3.4µs ± 0%    -88.42%  (p=0.000 n=10+9)
MinMaxUint64/2048KiB                233µs ± 0%        61µs ± 7%    -73.85%  (p=0.000 n=9+10)
MinMaxFloat32/4KiB                  819ns ± 0%        53ns ± 3%    -93.49%  (p=0.000 n=10+10)
MinMaxFloat32/256KiB               52.5µs ± 0%       3.7µs ± 2%    -92.87%  (p=0.000 n=8+10)
MinMaxFloat32/2048KiB               440µs ± 0%        62µs ± 6%    -85.83%  (p=0.000 n=10+10)
MinMaxFloat64/4KiB                  475ns ± 1%        54ns ± 2%    -88.71%  (p=0.000 n=10+10)
MinMaxFloat64/256KiB               29.2µs ± 0%       3.8µs ± 2%    -87.11%  (p=0.000 n=9+10)
MinMaxFloat64/2048KiB               235µs ± 0%        60µs ± 5%    -74.31%  (p=0.000 n=9+10)
MinMaxFixedLenByteArray/4KiB        544ns ± 0%       259ns ± 0%    -52.38%  (p=0.000 n=10+9)
MinMaxFixedLenByteArray/256KiB     34.1µs ± 0%      14.3µs ± 0%    -58.05%  (p=0.000 n=10+10)
MinMaxFixedLenByteArray/2048KiB     274µs ± 0%       155µs ± 5%    -43.54%  (p=0.000 n=10+10)

name                             old speed      new speed        delta
MinMaxBool/4KiB                  3.33GB/s ± 0%  103.77GB/s ± 0%  +3017.58%  (p=0.000 n=8+10)
MinMaxBool/256KiB                3.38GB/s ± 0%   79.14GB/s ± 0%  +2243.14%  (p=0.000 n=10+9)
MinMaxBool/2048KiB               3.36GB/s ± 0%   32.93GB/s ± 9%   +879.85%  (p=0.000 n=9+10)
MinMaxInt32/4KiB                 6.40GB/s ± 0%   92.97GB/s ± 0%  +1351.86%  (p=0.000 n=9+10)
MinMaxInt32/256KiB               6.44GB/s ± 0%   78.94GB/s ± 0%  +1125.56%  (p=0.000 n=10+9)
MinMaxInt32/2048KiB              6.42GB/s ± 0%   34.87GB/s ± 6%   +443.25%  (p=0.000 n=10+10)
MinMaxInt64/4KiB                 12.4GB/s ± 0%    91.0GB/s ± 0%   +635.50%  (p=0.000 n=9+8)
MinMaxInt64/256KiB               12.5GB/s ± 0%    77.7GB/s ± 0%   +519.81%  (p=0.000 n=10+9)
MinMaxInt64/2048KiB              12.4GB/s ± 0%    33.9GB/s ± 8%   +172.81%  (p=0.000 n=9+10)
MinMaxUint32/4KiB                4.53GB/s ± 0%   93.23GB/s ± 0%  +1958.77%  (p=0.000 n=8+10)
MinMaxUint32/256KiB              4.51GB/s ± 0%   79.20GB/s ± 0%  +1656.67%  (p=0.000 n=10+9)
MinMaxUint32/2048KiB             4.49GB/s ± 0%   35.48GB/s ± 7%   +689.68%  (p=0.000 n=10+10)
MinMaxUint64/4KiB                9.09GB/s ± 0%   89.22GB/s ± 0%   +881.23%  (p=0.000 n=10+8)
MinMaxUint64/256KiB              9.02GB/s ± 0%   77.87GB/s ± 0%   +763.67%  (p=0.000 n=10+8)
MinMaxUint64/2048KiB             8.99GB/s ± 0%   34.40GB/s ± 7%   +282.81%  (p=0.000 n=9+10)
MinMaxFloat32/4KiB               5.00GB/s ± 0%   76.92GB/s ± 3%  +1437.50%  (p=0.000 n=10+10)
MinMaxFloat32/256KiB             5.00GB/s ± 0%   70.05GB/s ± 2%  +1301.80%  (p=0.000 n=8+10)
MinMaxFloat32/2048KiB            4.77GB/s ± 0%   33.71GB/s ± 6%   +606.78%  (p=0.000 n=10+10)
MinMaxFloat64/4KiB               8.63GB/s ± 1%   76.43GB/s ± 2%   +786.13%  (p=0.000 n=10+10)
MinMaxFloat64/256KiB             8.97GB/s ± 0%   69.56GB/s ± 2%   +675.73%  (p=0.000 n=9+10)
MinMaxFloat64/2048KiB            8.94GB/s ± 0%   34.84GB/s ± 5%   +289.74%  (p=0.000 n=9+10)
MinMaxFixedLenByteArray/4KiB     7.53GB/s ± 0%   15.82GB/s ± 0%   +110.00%  (p=0.000 n=10+9)
MinMaxFixedLenByteArray/256KiB   7.69GB/s ± 0%   18.32GB/s ± 0%   +138.40%  (p=0.000 n=10+10)
MinMaxFixedLenByteArray/2048KiB  7.66GB/s ± 0%   13.59GB/s ± 5%    +77.30%  (p=0.000 n=10+10)
```