@achille-roussel 2022-02-08
> What's the use case for finding the max on 128 bit big endian? What Parquet data type does that correspond to?

I'm planning on using this for columns holding UUIDs. This would correspond to the FIXED_LEN_BYTE_ARRAY(16) type in parquet, with the UUID annotation https://github.com/apache/parquet-format/blob/master/LogicalTypes.md#uuid